% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-helpers.R
\name{agua_train_predict}
\alias{agua_train_predict}
\title{A helper function for processing a grid for a specific model.}
\usage{
agua_train_predict(static, grid, resample_label)
}
\arguments{
\item{static}{A list object produced by \pkg{tune}.}

\item{grid}{A tibble of candidate tuning parameters (if any)}

\item{resample_label}{A single row tibble of resampling \code{id} columns.}
}
\value{
A list of tibbles containing the out-of-sample predictions for each
row of \code{grid}.
}
\description{
This is an internal function that is used internal by the \pkg{tune}
package.
}
\keyword{internal}
