% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-ml.R
\name{rank_automl}
\alias{rank_automl}
\title{Tools for working with h2o auto_ml results}
\usage{
rank_automl(object, ...)
}
\arguments{
\item{object}{A \code{model_fit} or fitted \code{workflow} object.}

\item{n}{The number of individual models to extract from \code{auto_ml} results,
ranked descendingly by performance. Default to all.}

\item{id}{Model id.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of model's cross validation performances.
}
\description{
Tools for working with h2o auto_ml results
}
