% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.H2OAutoML}
\alias{autoplot.H2OAutoML}
\title{Plot the results of H2o AutoML}
\usage{
\method{autoplot}{H2OAutoML}(
  object,
  type = c("rank", "metric"),
  metric = NULL,
  std_errs = qnorm(0.95),
  ...
)
}
\arguments{
\item{object}{A fitted \code{auto_ml()} model.}

\item{type}{A single character of choices of "rank" (plotting average ranking
of algorithms within metrics) or "metric" (plotting average value of metrics).}

\item{metric}{A character vector or NULL for which metric to plot.
By default, all metrics will be shown via facets.}

\item{std_errs}{The number of standard errors to plot.}

\item{...}{Other options to pass to \code{autoplot()}.}
}
\value{
A ggplot object.
}
\description{
This \code{autoplot()} method plots cross validation performances of candidate
models in H2O AutoML output via facets on each metric.
}
\examples{
if (h2o_running()) {
  m <- auto_ml() \%>\%
    set_engine("h2o", max_runtime_secs = 10) \%>\%
    set_mode("regression") \%>\%
    fit(mpg ~ ., data = mtcars)

  autoplot(m)
}
}
