% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{h2o_start}
\alias{h2o_start}
\alias{h2o_end}
\alias{h2o_running}
\alias{h2o_remove}
\alias{h2o_remove_all}
\alias{h2o_get_model}
\alias{h2o_get_frame}
\alias{h2o_xgboost_available}
\title{Utility functions for interacting with the h2o server}
\usage{
h2o_start()

h2o_end()

h2o_running(verbose = FALSE)

h2o_remove(id)

h2o_remove_all()

h2o_get_model(id)

h2o_get_frame(id)

h2o_xgboost_available()
}
\arguments{
\item{verbose}{Print out the message if no cluster is available.}

\item{id}{Model or frame id.}
}
\description{
Utility functions for interacting with the h2o server
}
\examples{
if (!h2o_running()) {
  h2o_start()
}
}
\keyword{internal}
