% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h2o_running}
\alias{h2o_running}
\title{Check if h2o cluster is initialized}
\usage{
h2o_running(verbose = FALSE)
}
\arguments{
\item{verbose}{Print out the message if no cluster is available.}
}
\value{
A logical.
}
\description{
Check if h2o cluster is initialized
}
\examples{
h2o_running()
h2o_running(verbose = TRUE)
}
