% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h2o_start}
\alias{h2o_start}
\title{Silently start h2o server}
\usage{
h2o_start()
}
\value{
The output from \code{\link[h2o:h2o.init]{h2o::h2o.init()}}
}
\description{
Silently start h2o server
}
\examples{
h2o_start()
}
