% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{h2o_train}
\alias{h2o_train}
\alias{h2o_train_rf}
\title{Model wrappers for h2o}
\usage{
h2o_train(x, y, model, ...)

h2o_train_rf(x, y, ntrees = 50, mtries = -1, min_rows = 1, ...)
}
\arguments{
\item{x}{A data frame of predictors}

\item{y}{A vector of outcomes.}

\item{model}{A character string for the model.}

\item{...}{Other options to pass to the h2o model functions (e.g.,
\link[h2o:h2o.randomForest]{h2o::h2o.randomForest}).}

\item{ntrees}{Number of trees. Defaults to 50.}

\item{mtries}{Number of variables randomly sampled as candidates at each split. If set to -1, defaults to sqrt{p} for
classification and p/3 for regression (where p is the # of predictors Defaults to -1.}

\item{min_rows}{Fewest allowed (weighted) observations in a leaf. Defaults to 1.}
}
\value{
An h2o model object.
}
\description{
Basic model wrappers for h2o model functions that include data conversion,
seed configuration, and so on.
}
\examples{
# start with h2o::h2o.init()

if (h2o_running()) {
  spec <-
    rand_forest(mtry = 3, trees = 1000) \%>\%
    set_engine("h2o") \%>\%
    set_mode("regression")

  set.seed(1)
  mod <- fit(spec, mpg ~ ., data = mtcars)
  mod

  predict(mod, head(mtcars))

}
}
