% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r_h2o}
\alias{r_h2o}
\alias{h2o_r}
\title{Data conversion tools}
\usage{
r_h2o(x)

h2o_r(x)
}
\arguments{
\item{x}{A R data frame (for \code{r_h2o}) or an H2OFrame (\code{h2o_r}).}
}
\value{
A tibble or H2OFrame.
}
\description{
Data conversion tools
}
\examples{

# start with h2o::h2o.init()
if (h2o_running()) {
  cars2 <- r_h2o(mtcars)
  cars2

  cars0 <- h2o_r(cars2)
  cars0

}
}
