% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{\%>\%}
\alias{tidy}
\alias{refit}
\alias{autoplot}
\alias{extract_fit_parsnip}
\alias{extract_fit_engine}
\alias{collect_metrics}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{refit}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{hardhat}{\code{\link[hardhat:hardhat-extract]{extract_fit_engine}}, \code{\link[hardhat:hardhat-extract]{extract_fit_parsnip}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tune}{\code{\link[tune:collect_predictions]{collect_metrics}}}
}}

