% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua-package.R
\docType{package}
\name{agua-package}
\alias{agua}
\alias{agua-package}
\title{tidymodels integration with h2o}
\description{
agua has tools that allows users to fit and tune models using the h2o
platform. Most functions in this package are not user facing; agua enables
new parsnip model engines and sets up additional infrastructure for tune.
}
\details{
The package uses code initially written by Steven Pawley in his h2oparsnip
package. Addition work was done by Qiushi Yan as an RStudio summer intern.
}
\author{
\strong{Maintainer}: Qiushi Yan \email{qiushi.yan@rstudio.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Steven Pawley \email{dr.stevenpawley@gmail.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
