% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-ml.R
\name{h2o_rank}
\alias{h2o_rank}
\title{Tools for working with h2o auto_ml results}
\usage{
h2o_rank(object, ...)
}
\arguments{
\item{object}{A \code{model_fit} or fitted \code{workflow} object.}

\item{n}{The number of individual models to extract from \code{auto_ml} results,
ranked descendingly by performance.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of model's cross validation performances
}
\description{
Tools for working with h2o auto_ml results
}
