% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-ml.R
\name{rank_automl}
\alias{rank_automl}
\alias{rank_automl.default}
\alias{rank_automl.workflow}
\alias{rank_automl.model_fit}
\alias{tidy._H2OAutoML}
\alias{print.automl_fit}
\title{Tools for working with h2o auto_ml results}
\usage{
rank_automl(object, ...)

\method{rank_automl}{default}(object, ...)

\method{rank_automl}{workflow}(object, n = NULL)

\method{rank_automl}{model_fit}(object, n = NULL)

\method{tidy}{`_H2OAutoML`}(object, ...)

\method{print}{automl_fit}(object, ...)
}
\arguments{
\item{object}{A \code{model_fit} or fitted \code{workflow} object.}

\item{n}{The number of individual models to extract from \code{auto_ml} results,
ranked descendingly by performance. Default to all.}

\item{id}{Model id.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of model's cross validation performances.
}
\description{
\code{rank_automl()} returns a tibble ranking cross validation performances
of different algorithms on a metric. \code{tidy()} returns the leaderboard
in tidy format.
}
\examples{
if (h2o_running) {
  mod <- auto_ml() \%>\%
    set_engine("h2o", max_runtime_secs = 10) \%>\%
    set_mode("regression") \%>\%
    fit(mpg ~ ., data = mtcars)

  tidy(mod)
  rank_automl(mod)
}

}
