% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{h2o_activation}
\alias{h2o_activation}
\alias{h2o_split}
\title{Tuning parameters in h2o}
\usage{
h2o_activation(values = values_h2o_activation)

h2o_split(values = values_h2o_split)
}
\description{
Tuning parameters in h2o
}
\examples{
h2o_activation()
}
\keyword{internal}
