% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{h2o_activation}
\alias{h2o_activation}
\title{Tuning parameters in h2o}
\usage{
h2o_activation(
  values = c("Rectifier", "RectifierWithDropout", "Tanh", "TanhWithDropout", "Maxout",
    "MaxoutWithDropout")
)
}
\description{
Tuning parameters in h2o
}
\keyword{internal}
